third_party/blink/web_tests/svg/custom/disallow-non-lengths-in-attrs-expected.txt

CONSOLE ERROR: Error: <svg> attribute width: Expected length, "auto".
CONSOLE ERROR: Error: <foreignObject> attribute width: Expected length, "auto".
CONSOLE ERROR: Error: <svg> attribute width: Expected length, "initial".
CONSOLE ERROR: Error: <foreignObject> attribute width: Expected length, "initial".
CONSOLE ERROR: Error: <svg> attribute width: Expected length, "inherit".
CONSOLE ERROR: Error: <foreignObject> attribute width: Expected length, "inherit".
CONSOLE ERROR: Error: <svg> attribute width: Expected length, "foo".
CONSOLE ERROR: Error: <foreignObject> attribute width: Expected length, "foo".
This is a testharness.js-based test.
[PASS] Test width 'auto' on SVGSVGElement
[PASS] Test width 'auto' on SVGForeignObject
[PASS] Test width 'initial' on SVGSVGElement
[PASS] Test width 'initial' on SVGForeignObject
[PASS] Test width 'inherit' on SVGSVGElement
[PASS] Test width 'inherit' on SVGForeignObject
[PASS] Test width 'foo' on SVGSVGElement
[PASS] Test width 'foo' on SVGForeignObject
Harness: the test ran to completion.
